First <provider> always wins.
In scanPackageLI(), the first provider to claim an authority wins
over any other later providers.  To match this behavior, this changes
ProviderIntentResolver to ignore later providers with the same
ComponentName.

Without this behavior, AM.getContentProvider() could end up wedged
waiting for a provider that generateApplicationProvidersLocked()
never decided to spin up.

Bug: 12956224
Change-Id: I37d3e6e9cf3eea2cac4c0e914134f6c83b2fc2bf

Former-commit-id: c6809bb6853c6f055905cbbbf873bbe4066c5d21