Support for new gTLDs
Bug: 13006774

Add support for new generic Top-Level Domains (gTLDs) from ICANN
and deprecate the public APIs that depend on them. Due to the
fast profileration of gTLDs, it has become hard to keep this API
up-to-date. Further it already fails to recognize internal domains.
We thus replace them with simple "any letter sequence" regexps.

This CL is almost same as  Ib6cb5360bbb4a02b14bb34acb30d1cc8ddec771b.
The difference is that the values of the public gTLD strings are only
deprecated and not modified.

Change-Id: I37c130ace9068b80eeaec3282a77450517ed7794

Former-commit-id: b080c3c54a79493863aee3f4ba5d96307ac500ec