Remove call to setIsOpaque (again).
The call was removed in I1b36b0b0ce7126031eb7b769b563c17dcd4b306a,
since setIsOpaque was deprecated in Skia.

It was accidentally reintroduced in
Id5793fa0ebc17ee8b1eecf4b3f327977fdccff71. This line is redundant
with the lines above it (which call the replacement for setIsOpaque,
setAlphaType).

Change-Id: I6520f796d639721987418ec0ecae82b3674dcbfd

Former-commit-id: 73c04711ca13c99f29a84de8cfc6c780811a8f63