Allow a user to elect to see private notifications in public.
Settings.Secure.LOCK_SCREEN_ALLOW_PRIVATE_NOTIFICATIONS, if
set and nonzero, will show that user's notifications on the
secure lockscreen (assuming LOCK_SCREEN_ALLOW_NOTIFICATIONS
is also set). Note that this applies only to a particular
user's notifications, so if multiple users' notifications
are shown at once (currently this only applies to the
current user and USER_ALL) only that user's notifications
are affected.

USER_ALL is assumed to always allow private notifications in
public (because they can't be user-specific).

Change-Id: I32bb7939066c4919191e74161ca48ead01a88c7f

Former-commit-id: fd16d56419649b885f2e18dc37a18add0fe91e01