Prevent authenticators from using Settings to  launch arbitrary activities.
Various authenticator results such as getAuthToken and addAccount might
result in an Intent returned to the AccountManager caller. A malicious
authenticator could exploit the fact that the Settings are a system app,
lead the user to launch add account for their account type and thus get
Settings to use the intent to start some arbitrary third parties Activity.

The fix is to make sure that the UID of the app associated with Activity
to be launched by the supplied intent and the Authenticators UID share
the same signature.  This means that an authenticator implementer can only
exploit apps they control.

This is a backport of a53e8c40f5a66c0a0983bd0c3d83f8d8206a0a27 [formerly 5bab9daf3cf66f4de19f8757e386030e8bef23ce]

Bug: 7699048
Change-Id: Ifed345c2fc20020d55fa2cab1f2f7ea509ea09b2

Former-commit-id: b9ba0c6c43f0f7f0d5e347030e3c86a86ed74542