Extend DeviceAdminReceiver to receive provisioning complete broadcast.
The managed profile provisioning app completes provisioning (creating the profile, installing the mdm on the secondary user, setting the mdm as the profile owner, removing unneeded apps from secondary user, removing mdm from the primary user). Then it sends a notification that the provisioning has completed across to the new profile. If the mdm, which is now installed on the profile wants to be notified that provisioning has completed it needs to implement this callback method.

Change-Id: I75f2d6d19709fd39aa867f1254e0c37a0c936222

Former-commit-id: 8cdb6fcd9b6a53978cb2b2ad9ab668ae19392266