Allow click listener registration in NumberPicker.
NumberPicker is composed of three areas, increment, decrement, and value,
which take the entire space. Hence, adding a click listener was a no-op.
Clicking on the value brings up the IME but for devices with very small
screens a developer may want to override this default behavior to say
confirm the selection.

This change allows adding a click listener to NumberPicker to override the
behavior of clicking on the current value. This is applicable only to the
new look and feel of the NumberPicker, i.e. the holo themes. This change
is safe as previously setting a click listener had no effect.

bug:13287234

Change-Id: I34e12a2e2bd64344a4797153fa6c820001a4722e
(cherry picked from commit bf9898e9e0bd5d0ae0fa4c5a91b150666b29bf33 [formerly 46a27efe9528ac3c5b37726707844e7c868815ac])

Former-commit-id: 1caa092f94aaa4b9b7b382263dfd13581d92d2de