Fix log spam from View.toString().
If a view has a non-zero ID that is not defined in resources (i.e. has been
set pragmatically), the calls to Resources.getResourcePackageName() and
Resources.getResourceEntryName() result in a log warning:
No package identifier when getting name for resource number 0x00000001

Fix this by not attempting to resolve the package & name when there is
none.

Change-Id: Id88a61539fffb36187da7911f8e8a42d5a1bb951

Former-commit-id: d15990aa501b683c143700e246b390c8bb79b210