Activity chooser "Always" button is not working
When ResolverActivity is created with a custom list of matching
applications (rList) as in NFC case, and the alwaysUseOption is
set to true, the prferredActivity is not saved even if the user
presses the "always" button.

When a list is provided the variable mBaseResolveList will be
!= null. This will set mOrigResolveList = null.

When an activity is choosen and one of the buttons are pressed
onIntentSelected is called. The first thing this method does
is to check mAdapter.mOrigResolveList != null, however in this
case mOrigResolveList is always null, and the value is not
saved as PreferredActivity.

This problem was introduced in
1b696b7b69496c1ab478a9fc5a0ad8baae3b73c3 [formerly 6d8dfbd8149942f25f2b9643a12f1fb38f3be834].

Change-Id: I9eac41b7861b5e68ad3978af0dc0285f2a34eb88

Former-commit-id: e88b042d1d859ec754ed5a5f366bb560b30f67ec