BootReceiver: add ro.boot.bootreason property to SYSTEM_LAST_KMSG
Instead of requiring the kernel to join last_kmsg and the reboot
reason together, allow both to be passed to userspace and have
userspace combine them the way the log parsers want them.

Existing devices with no ro.boot.bootreason property and kernel
support for putting the reason in last_kmsg will continue to
use the kernel's formatting.

Bug: 13813279
Change-Id: I079b0107feb1533c6a54044ca6a114741127dfbc

Former-commit-id: 431614cb6b2170fc715757dc6f964fdbb140ee7d