Flatten notification hiearchy and remove glow.
Currently, the padding and the glow was inside the individual
notification. This no longer works if we want to adjust the
padding dynamically whether we are on Keyguard or not. This change
moves the padding outside of the individual notifications, and as
a side effect, removes the glow. The glow wasn't really visible with
the new layout, so it's not a breaking change. We have to discuss
with UX first what the new "glow" solution is going to be.

Change-Id: Iac16892cb3b7bb1de3001954b1428796b07950c1

Former-commit-id: fe40f7d13bfc1faa35c9a131ce4be5104cb8f6b9