Fix Global Proxy when used with PAC
Since PAC needs to relay the local proxy port back to the
ConnectivityService it ends up calling handleApplyDefaultProxy...
This works fine for PAC on WiFi, but when tested on global proxy
(not currently used anywhere), it sets the mDefaultProxy.  This
mDefaultProxy does not get cleared when the global proxy is cleared
and requires a reboot to get things cleared out.

This CL adds a check to overwrite mGlobalProxy rather than
mDefaultProxy in this use case.

Change-Id: I92782d11e213b91f8ddda2faaf996a7252273fc3

Former-commit-id: 56cf1ab4105cffd5afb3abd1411fd38d98002853