Fixed several bugs regarding the state of the notifications
Fixed an issue where the last card was hidden even though there would
have been enough space to show it when it was expanded.
Also fixed an issue where the first card was not expandable when it was
the only one.
In addition the scroll range was corrected, to ensure that the last card
could always be scrolled to full visibility, which was not the case
previously when the last card was expanded.

Bug: 14376035
Change-Id: I372a55dc1fc6e381dd9aef3d631447912e3d1024

Former-commit-id: 4a1ac84e35fe5a8b1f1de01193e51da173eea99c