Fix mismatch in Transition ID/instances
Bug 14597573

In a Transition, when a View is removed from a Scene
and that View has an ID in common with another View
in the end Scene, it will be matched, even if the
View in the start Scene already had a matching View
by instance. This CL prevents a match by ID when an
instance match is available.

Future: disable matching when two views have the
same ID.

Change-Id: I0e51d1f2d0a2d05d09e8e135f090d0e1b2a67efc

Former-commit-id: 6de56bb07ec58480983932cb055aa9ed2ddfcc99