New, inkier EdgeEffect visual style
Remove the two-part gradient and edge glow and replace it with a
circular ink-like style instead. This now follows the theme color
palette so that it can be more in tune with the app's overall visual
design.

Add API to EdgeEffect to specify a horizontal displacement for the
effect. This lets the larger middle section follow relevant touch
points.

Fix a bug that caused EdgeEffect pulls to stutter in ListView/GridView
as a result of some annoying state-fluttering.

Change-Id: Ief144b06bd9d3dbe85d3154585a703fa552b0808

Former-commit-id: c501db9f44f7967961f0ba61a0b3b63055ac1190