Adding shell command execution to UiAutomation.
The UiAUtomation APIs are an extension object on instrumentation and
is not null only if the instrumentation was started from the shell.
The UiAutomation APIs allow performing privileged operations for testing.
Since UiAutomation tests are just regular instrumentation tests they
cannot take advantage of the rich set of command line utilities on the
platform. This change adds a capability to UiAutomation to execute
shell commands.

bug:12927198

Change-Id: Ifb764c8e89943654f3e35d7a0c65b0b730db672f

Former-commit-id: 121e0c073992658ca0ba055f40bf3b130caa819a