Prevent home task from being set as ontopofhome
If the front activity of a task that is on top of home finishes, the oldest task in
other stacks will be set as ontopofhome (which in this case meant the
home task was set to ontopofhome).  The solution is to ensure we
only set a task to ontopofhome if the finishing activity is in the same
stack.

Change-Id: Idd8f9af4d3d9bd2de3c519d3dcc60f479695ae0c

Former-commit-id: ccd230712d8ec4c017f8ab9b24ee53b4e5dfc19b