Implementation of TaskManager reschedule/cancel
Issues here:
"Reschedule" of an idle-mode task is not well-defined. In the
API I throw an error if you try to set a back-off policy on
an idle mode task.
Implementation-wise, i add a delay for a reschedule request of an
idle mode task. This means that if the phone's still in idle mode
after the delay they app will get a call back, but otherwise it'll
have to wait til the next one.

Implemented all API functions

Change-Id: I0264c2614cc52c28730abbcb1557f314f058f991

Former-commit-id: b61c506a5002d7cf3bae535bfde919a0c72cd8c0