DO NOT MERGE Implementation of TaskManager reschedule/cancel
Issues here:
"Reschedule" of an idle-mode task is not well-defined. In the
API I throw an error if you try to set a back-off policy on
an idle mode task.
Implementation-wise, i add a delay for a reschedule request of an
idle mode task. This means that if the phone's still in idle mode
after the delay they app will get a call back, but otherwise it'll
have to wait til the next one.

Implemented all API functions

Former-commit-id: 9b9244b6941110ea2d940d9fc8eed0cdff96a016