Distinguish between menus in ActionBar and popup. [DO NOT MERGE]
Mark the menus in the ActionBar and the ones in the Overflow popup
separately. This enables the IDE to differentiate between the two and
enable gestures such as double-click to open the menu xml for menus in
the ActionBar and single click for the menus in the Overflow popup.

Change-Id: I7a23f87d34d618a134ffbe3a665be695915d4bae
(cherry picked from commit 8247b46c89fa0340405ba0839683435f8f2851b9 [formerly 03c0735017a34bac9ea69600b8cf980775331f0e])

Former-commit-id: 3ff08eb6bf68868d4f4e872fbbf1ecaf521934a9