Distinguish between menus in ActionBar and popup. [DO NOT MERGE]
Mark the menus in the ActionBar and the ones in the Overflow popup
separately. This enables the IDE to differentiate between the two and
enable gestures such as double-click to open the menu xml for menus in
the ActionBar and single click for the menus in the Overflow popup.

Change-Id: I7a23f87d34d618a134ffbe3a665be695915d4bae
(cherry picked from commit bb359359187ba6acec5b9a7c4859577cf02b6243 [formerly 866d40fc39f7364d95085650244db59f52d4c002])

Former-commit-id: 03c0735017a34bac9ea69600b8cf980775331f0e