Distinguish between menus in ActionBar and popup. [DO NOT MERGE]
Mark the menus in the ActionBar and the ones in the Overflow popup
separately. This enables the IDE to differentiate between the two and
enable gestures such as double-click to open the menu xml for menus in
the ActionBar and single click for the menus in the Overflow popup.

Change-Id: I7a23f87d34d618a134ffbe3a665be695915d4bae
(cherry picked from commit 41ee4f6970ce218e060429fd83d1d85da76da502 [formerly 79f45b43db4dfd605d1e7ddd2c71be4ef8285269])

Former-commit-id: 866d40fc39f7364d95085650244db59f52d4c002