Update InputMethodSubtypeSwitchingController correctly
User visible behavior is supposed to be identical with and
without this CL.

Previously, there is some corner cases where
InputMethodSubtypeSwitchingController#resetCircularListLocked is
not called but the list of enabled input method is updated.

Fortunately, this corner cases are not observable for a user
because we have not updated the the rotation order dynamically.

However we should fix this before implementing smarter rotation
algorithm that changes the rotation order dynamically.

BUG: 7043015
Change-Id: I145a514dc4cde369ba50431c408c916046ab0c6e

Former-commit-id: c834a2590cc7ac478ba2ef5a6d8eb7ce471df132