Add lock-to-app mode
Added a dialog that shows when app has not been authorized by
DevicePolicyManager.isLockTaskAuthorized.  This allows any app
to trigger lock-to-app mode.  This same dialog is used when
startLockTaskOnCurrent is triggered by the recents long-press.

Can exit the mode by long-pressing recents again.

Keyguard is disabled when lock-to-app is active.

This CL also prevents apps from finishing when they are the root
task in a lock task TaskRecord.

Change-Id: Ib54d858e570cccf6bfd986958868e15f49bcef75

Former-commit-id: 62515beee67307d8859beec521b7baedfa54b2b5